+++ src/main/java/roidrole/roidtweaker/RoidTweaker.java
@@ -5,8 +5,8 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import rocks.gameonthe.rockytweaks.crafttweaker.merchant.MerchantCommand;
 import roidrole.roidtweaker.mods.minecraft.anvil.AnvilListener;
+import roidrole.roidtweaker.mods.minecraft.villager.VillagerCommand;
 import xyz.tcreopargh.ctintegration.CTIntegration;
 
 @Mod(
@@ -32,6 +32,6 @@ public class RoidTweaker {
     @Mod.EventHandler
     public void init(FMLInitializationEvent event) {
         MinecraftForge.EVENT_BUS.register(new AnvilListener());
-        CTChatCommand.registerCommand(new MerchantCommand());
+        CTChatCommand.registerCommand(new VillagerCommand());
     }
 }