+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 -
 First release
 
+Note : The listed changes for this version are relative to the included mods
+
 **Added**: 
 - CraftTweaker Integration
-- Rocky Tweaks
+- Rocky Tweaks' functionality (rewritten, new namespace)
+- Vanilla Anvil Repair's CT functionality (rewritten, new namespace)
 - Config handling
 - ProjectE custom fuels
 
@@ -14,5 +17,4 @@ First release
 - Crash with client.player.difficulty (CraftTweaker Integration)
 
 **Changed**:
-- Slight optimization to RecipePattern.build();
-- Rewrote Anvil recipe handling entirely
+- Slight optimization to RecipePattern.build(); (CraftTweaker Integration)