+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.1.1 : 
+Fix mod not loading
 
 ## 1.0.0 :
 First release